sync tests fixed + worker tests fixed
