+++ backend/Dockerfile.dev
@@ -36,7 +36,9 @@ COPY worker/quivr_worker/__init__.py ./worker/quivr_worker/__init__.py
 
 RUN PYTHONDONTWRITEBYTECODE=1 pip install --no-cache-dir -r requirements.lock
 
-RUN playwright install --with-deps
+RUN playwright install --with-deps && \
+    python -c "from unstructured.nlp.tokenize import download_nltk_packages; download_nltk_packages()" && \
+    python -c "import nltk;nltk.download('punkt_tab'); nltk.download('averaged_perceptron_tagger_eng')"
 
 COPY . .
