+++ backend/pyproject.toml
@@ -88,3 +88,13 @@ section-order = [
     "local-folder",
 ]
 known-first-party = []
+
+[tool.pytest.ini_options]
+addopts = "--tb=short -ra -v"
+filterwarnings = ["ignore::DeprecationWarning"]
+markers = [
+    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
+    "base: these tests require quivr-core with extra `base` to be installed",
+    "tika: these tests require a tika server to be running",
+    "unstructured: these tests require `unstructured` dependency",
+]
